+++ src/main/java/cn/nukkit/network/protocol/LevelEventGenericPacket.java
@@ -10,7 +10,7 @@ import java.nio.ByteOrder;
 @ToString
 public class LevelEventGenericPacket extends DataPacket {
 
-    public static final byte NETWORK_ID = ProtocolInfo.LEVEL_EVENT_GENERIC_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.LEVEL_EVENT_GENERIC_PACKET;
 
     public int eventId;
     public CompoundTag tag;
