+++ src/main/java/cn/nukkit/network/protocol/SetPlayerGameTypePacket.java
@@ -8,7 +8,7 @@ import lombok.ToString;
  */
 @ToString
 public class SetPlayerGameTypePacket extends DataPacket {
-    public final static byte NETWORK_ID = ProtocolInfo.SET_PLAYER_GAME_TYPE_PACKET;
+    public final static int NETWORK_ID = ProtocolInfo.SET_PLAYER_GAME_TYPE_PACKET;
 
     @Override
     public int pid() {
